+++ examples/removeDlmmLiquidityAndZapOut.ts
@@ -11,6 +11,7 @@ import { getJupiterQuote, getJupiterSwapInstruction } from "../src/helpers";
 import DLMM, { getTokenProgramId } from "@meteora-ag/dlmm";
 import { DLMM_PROGRAM_ID } from "../src";
 import { TOKEN_PROGRAM_ID } from "@solana/spl-token";
+import { JUPITER_API_KEY, JUPITER_API_URL } from "./constants";
 
 async function main() {
   const connection = new Connection("https://api.mainnet-beta.solana.com");
@@ -29,7 +30,10 @@ async function main() {
     "9NRifL3nKQU84hMTbhE7spakkGy5vq4AvNHNQYr8LkW7"
   );
 
-  const zap = new Zap(connection);
+  const zap = new Zap(connection, {
+    jupiterApiUrl: JUPITER_API_URL,
+    jupiterApiKey: JUPITER_API_KEY,
+  });
   const dlmm = await DLMM.create(connection, lbPairAddress, {
     cluster: "mainnet-beta",
     programId: new PublicKey(DLMM_PROGRAM_ID),
@@ -124,7 +128,10 @@ async function main() {
         false,
         true,
         true,
-        "https://lite-api.jup.ag"
+        {
+          jupiterApiUrl: JUPITER_API_URL,
+          jupiterApiKey: JUPITER_API_KEY,
+        }
       ),
     ]);
 
@@ -196,7 +203,11 @@ async function main() {
     } else if (bestProtocol === "jupiter" && quotes.jupiter) {
       const swapInstructionResponse = await getJupiterSwapInstruction(
         wallet.publicKey,
-        quotes.jupiter
+        quotes.jupiter,
+        {
+          jupiterApiUrl: JUPITER_API_URL,
+          jupiterApiKey: JUPITER_API_KEY,
+        }
       );
 
       zapOutTx = await zap.zapOutThroughJupiter({