+++ examples/zapInDlmmIndirectSingleSided.ts
@@ -11,6 +11,7 @@ import Decimal from "decimal.js";
 import { NATIVE_MINT } from "@solana/spl-token";
 import BN from "bn.js";
 import { createJitoTipIx, getKeypairFromSeed, sendJitoBundle } from "./helpers";
+import { JUPITER_API_KEY, JUPITER_API_URL } from "./constants";
 
 const JITO_PRIVATE_KEY = process.env.JITO_PRIVATE_KEY!;
 
@@ -30,7 +31,10 @@ async function main() {
 
   const amountUseToAddLiquidity = new BN(0.001 * LAMPORTS_PER_SOL);
 
-  const zap = new Zap(connection);
+  const zap = new Zap(connection, {
+    jupiterApiUrl: JUPITER_API_URL,
+    jupiterApiKey: JUPITER_API_KEY,
+  });
   const dlmm = await DLMM.create(connection, dlmmPool);
   const binDelta = 34;
   const singleSided = DlmmSingleSided.X; // MET
@@ -51,6 +55,10 @@ async function main() {
     maxDeltaId,
     strategy: StrategyType.Spot,
     singleSided,
+    config: {
+      jupiterApiUrl: JUPITER_API_URL,
+      jupiterApiKey: JUPITER_API_KEY,
+    },
   });
 
   const result = await zap.getZapInDlmmIndirectParams({