+++ src/types.ts
@@ -17,6 +17,11 @@ import {
 
 export type ZapProgram = Program<Zap>;
 
+export type ZapConfig = {
+  jupiterApiUrl?: string;
+  jupiterApiKey?: string;
+};
+
 ///// ZAPOUT TYPES /////
 export type ZapOutParameters = IdlTypes<Zap>["zapOutParameters"];
 
@@ -270,6 +275,7 @@ export interface EstimateDlmmDirectSwapParams {
   maxDeltaId: number;
   strategy: StrategyType;
   singleSided?: DlmmSingleSided;
+  config?: ZapConfig;
 }
 
 export interface DlmmDirectSwapEstimateContext {
@@ -305,6 +311,7 @@ export interface EstimateDlmmRebalanceSwapParams {
   maxDeltaId: number;
   swapSlippageBps: number;
   strategy: StrategyType;
+  config?: ZapConfig;
 }
 
 export interface DlmmDirectRebalanceEstimateContext {
@@ -332,6 +339,7 @@ export interface EstimateDlmmIndirectSwapParams {
   maxDeltaId: number;
   strategy: StrategyType;
   singleSided?: DlmmSingleSided;
+  config?: ZapConfig;
 }
 
 export interface DlmmIndirectSwapEstimateResult {